feat: add support for dts.config.ts
attempts to load dts.config.ts if it is available,
throws an error if unable to load it. if dts.config.ts
isn't available, moves on to check for dts.config.js.

enables type declarations in tsconfig.json so that
people can import types correctly.

this commit does not include documentation about how
to use dts.config.ts as type exports should probably
be revisited.

closes #128